finishing extended functions tests & corrections

+++ extends.js
@@ -451,14 +451,33 @@ const path = require('path'), fs = require('fs');
 		else if ('string' !== typeof target) {
 			throw new Error("'target' is not a string");
 		}
+		else {
 
-		if (fs.isFileSync(target)) {
-			fs.unlinkSync(target);
+			origin = origin.trim();
+			target = target.trim();
+
+			if ('' === origin) {
+				throw new Error("'origin' is empty");
+			}
+				else if (!fs.isFileSync(origin)) {
+					throw new Error("'origin' is not a file");
+				}
+			else if ('' === target) {
+				throw new Error("'target' is empty");
+			}
+			else {
+
+				if (fs.isFileSync(target)) {
+					fs.unlinkSync(target);
+				}
+
+				fs.writeFileSync(target, fs.readFileSync(origin));
+
+				return true;
+
+			}
+
 		}
-
-		fs.writeFileSync(target, fs.readFileSync(origin));
-
-		return true;
 
 	};
 
@@ -504,30 +523,57 @@ const path = require('path'), fs = require('fs');
 			}
 			else {
 
-				fs.isFile(target, function(err, exists) {
+				origin = origin.trim();
+				target = target.trim();
 
-					if (err) {
-						callback(err);
-					}
-					else if (!exists) {
-						_copy(origin, target, callback);
-					}
-					else {
+				if ('' === origin) {
+					callback("'origin' is empty");
+				}
+				else if ('' === target) {
+					callback("'target' is empty");
+				}
+				else {
 
-						fs.unlink(target, function(err) {
+					fs.isFile(origin, function(err, exists) {
 
-							if (err) {
-								callback((err.message) ? err.message : err);
-							}
-							else {
-								_copy(origin, target, callback);
-							}
+						if (err) {
+							callback(err);
+						}
+						else if (!exists) {
+							callback("'origin' is not a file");
+						}
+						else {
 
-						});
+							fs.isFile(target, function(err, exists) {
 
-					}
+								if (err) {
+									callback(err);
+								}
+								else if (!exists) {
+									_copy(origin, target, callback);
+								}
+								else {
 
-				});
+									fs.unlink(target, function(err) {
+
+										if (err) {
+											callback((err.message) ? err.message : err);
+										}
+										else {
+											_copy(origin, target, callback);
+										}
+
+									});
+
+								}
+
+							});
+					
+						}
+
+					});
+
+				}
 
 			}
